requirements.txt.txt
flask
flask-ask
pyserial
requests